Enum ataxx::Square

pub enum Square {

    A1,
    B1,
    C1,
    D1,
    E1,
    F1,
    G1,
    A2,
    B2,
    C2,
    D2,
    E2,
    F2,
    G2,
    A3,
    B3,
    C3,
    D3,
    E3,
    F3,
    G3,
    A4,
    B4,
    C4,
    D4,
    E4,
    F4,
    G4,
    A5,
    B5,
    C5,
    D5,
    E5,
    F5,
    G5,
    A6,
    B6,
    C6,
    D6,
    E6,
    F6,
    G6,
    A7,
    B7,
    C7,
    D7,
    E7,
    F7,
    G7,

}

Square represents all the squares present on an Ataxx Board. The index of each Square is equal to rank-index * 8 + file-index.

Variants

A1

B1

C1

D1

E1

F1

G1

A2

B2

C2

D2

E2

F2

G2

A3

B3

C3

D3

E3

F3

G3

A4

B4

C4

D4

E4

F4

G4

A5

B5

C5

D5

E5

F5

G5

A6

B6

C6

D6

E6

F6

G6

A7

B7

C7

D7

E7

F7

G7

Implementations

impl Square

pub const N: usize = 49usize

N represents the total number of Squares in an Ataxx Board.

pub fn new(file: File, rank: Rank) -> Square

new creates a new Square from the given File and Rank.

use ataxx::*;

assert_eq!(Square::new(File::A, Rank::First), Square::A1);

pub fn file(self) -> File

file returns the File of the current Square.

use ataxx::*;

assert_eq!(Square::A1.file(), File::A);

pub fn rank(self) -> Rank

rank returns the Rank of the current Square.

use ataxx::*;

assert_eq!(Square::A1.rank(), Rank::First);

pub fn north(self) -> Self

north returns the Square to the North of this one.

use ataxx::*;

assert_eq!(Square::D4.north(), Square::D5);

pub fn south(self) -> Self

south returns the Square to the South of this one.

use ataxx::*;

assert_eq!(Square::D4.south(), Square::D3);

pub fn east(self) -> Self

east returns the Square to the East of this one.

use ataxx::*;

assert_eq!(Square::D4.east(), Square::E4);

pub fn west(self) -> Self

west returns the Square to the West of this one.

use ataxx::*;

assert_eq!(Square::D4.west(), Square::C4);

pub fn unsafe_from<T: ToPrimitive>(number: T) -> Self