Struct ataxx::MoveList

pub struct MoveList { /* private fields */ }

MoveList is a basic implementation of MoveStore that is used to allow users to utilize move-generation methods without having to implement a MoveStore by themselves. It also has utility methods other than the MoveStore trait.

Implementations

impl MoveList

pub fn new() -> MoveList

new creates an empty MoveList ready for use.

use ataxx::*;

let movelist = MoveList::new();

assert!(movelist.is_empty());

pub fn at(&self, n: usize) -> Move

at returns the Move stored at the given index. This operation is defined and valid if and only if the 0 <= index <= MoveList length.