Struct ataxx::Position

pub struct Position {
    pub bitboards: [BitBoard; 3],
    pub checksum: Hash,
    pub side_to_move: Color,
    pub ply_count: u16,
    pub half_move_clock: u8,
}

Position represents the snapshot of an Ataxx Board, the state of the an ataxx game at a single point in time. It also provides all of the methods necessary to manipulate such a snapshot.

Fields
bitboards: [BitBoard; 3]

bitboards stores BitBoards for the piece configuration of each Piece.

checksum: Hash

checksum stores the semi-unique Hash of the current Position.

side_to_move: Color

side_to_move stores the Piece whose turn to move it currently is.

ply_count: u16half_move_clock: u8

half-move clock stores the number of half-moves since the last irreversible Move. It is used to adjudicate games using the 50-move/100-ply rule.

Implementations

impl Position

pub fn new( black: BitBoard, white: BitBoard, block: BitBoard, side_to_move: Color, ply_count: u16, half_move_clock: u8, ) -> Position

new creates a new Position with the given BitBoards and side to move.

pub fn put(&mut self, sq: Square, piece: Piece)

put puts the given piece represented by its Piece on the given Square.

use ataxx::*;

let mut position = Position::new(
    BitBoard::EMPTY,
    BitBoard::EMPTY,
    BitBoard::EMPTY,
    Color::Black,
    0, 0
);
position.put(Square::A1, Piece::White);

assert_eq!(position.color_bb(Color::White), bitboard! {
    . . . . . . .
    . . . . . . .
    . . . . . . .
    . . . . . . .
    . . . . . . .
    . . . . . . .
    X . . . . . .
});

pub const fn at(&self, sq: Square) -> Option<Piece>

at returns the Piece of the piece present on the given Square.

use ataxx::*;

let position = Position::new(
    BitBoard::UNIVERSE,
    BitBoard::EMPTY,
    BitBoard::EMPTY,
    Color::Black,
    0, 0
);
assert_eq!(position.at(Square::A1), Some(Piece::Black));

pub const fn piece_bb(&self, piece: Piece) -> BitBoard

bitboard returns the BitBoard associated to the piece configuration of the given Piece. Only the Squares with a piece of the given Piece on them are contained inside the returned BitBoard.

use ataxx::*;

let position = Position::new(
    BitBoard::UNIVERSE,
    BitBoard::EMPTY,
    BitBoard::EMPTY,
    Color::Black,
    0, 0
);
assert_eq!(position.color_bb(Color::Black), BitBoard::UNIVERSE);

pub const fn color_bb(&self, color: Color) -> BitBoard

impl Position

pub fn is_game_over(&self) -> bool

is_game_over checks if the game is over, i.e. is a win or a draw.

use ataxx::*;
use std::str::FromStr;

let black_win = Position::from_str("xxxxxxx/7/7/7/7/7/7 o 0 1").unwrap();
let white_win = Position::from_str("ooooooo/7/7/7/7/7/7 x 0 1").unwrap();
let draw_game = Position::from_str("xxx1ooo/7/7/7/7/7/7 x 100 1").unwrap();
let ongoing = Position::from_str("xxx1ooo/7/7/7/7/7/7 x 0 1").unwrap();

assert!(black_win.is_game_over());
assert!(white_win.is_game_over());
assert!(draw_game.is_game_over());
assert!(!ongoing.is_game_over());

pub fn winner(&self) -> Option<Piece>

winner returns the Piece which has won the game. It returns None if the game is a draw. If Position::is_game_over is false, then the behavior of this function is undefined.

use ataxx::*;
use std::str::FromStr;

let black_win = Position::from_str("xxxxxxx/7/7/7/7/7/7 o 0 1").unwrap();
let white_win = Position::from_str("ooooooo/7/7/7/7/7/7 x 0 1").unwrap();
let draw_game = Position::from_str("xxx1ooo/7/7/7/7/7/7 x 100 1").unwrap();

assert_eq!(black_win.winner(), Some(Piece::Black));
assert_eq!(white_win.winner(), Some(Piece::White));
assert_eq!(draw_game.winner(), None);

impl Position

pub fn after_move<const UPDATE_HASH: bool>(&self, m: Move) -> Position

after_move returns a new Position which occurs when the given Move is played on the current Position. Its behavior is undefined if the given Move is illegal.

use ataxx::*;
use std::str::FromStr;

let mut pos = Position::from_str("x5o/7/7/7/7/7/o5x x 0 1").unwrap();
let new_pos = Position::from_str("xx4o/7/7/7/7/7/o5x o 0 1").unwrap();

let mov = Move::new_single(Square::B7);

assert_eq!(pos.after_move::<true>(mov).checksum, new_pos.checksum);

impl Position

pub fn generate_moves(&self) -> MoveList

generate_moves generates the legal moves in the current Position and returns a MoveList containing all the moves. It is a wrapper on top of the more general Position::generate_moves_into<T>.

use ataxx::*;
use std::str::FromStr;

let position = Position::from_str("x5o/7/7/7/7/7/o5x x 0 1").unwrap();
let movelist = position.generate_moves();

// There are 16 possible moves in startpos.
assert_eq!(movelist.len(), 16);

pub fn generate_moves_into<T: MoveStore>(&self, movelist: &mut T)

generate_moves_into generates all the legal moves in the current Position and adds them to the given movelist. The type of the movelist must implement the MoveStore trait.

use ataxx::*;
use std::str::FromStr;

let position = Position::from_str("x5o/7/7/7/7/7/o5x x 0 1").unwrap();
let mut movelist = MoveList::new();

position.generate_moves_into(&mut movelist);

// There are 16 possible moves in startpos.
assert_eq!(movelist.len(), 16);

pub fn count_moves(&self) -> usize

count_moves returns the number of legal moves in the current Position. It is faster than calling Position::generate_moves or Position::generate_moves_into<T> and then finding the length.

use ataxx::*;
use std::str::FromStr;

let position = Position::from_str("x5o/7/7/7/7/7/o5x x 0 1").unwrap();

// There are 16 possible moves in startpos.
assert_eq!(position.count_moves(), 16);