Trait ataxx::MoveStore

pub trait MoveStore {
    // Required methods
    fn push(&mut self, m: Move);
    fn len(&self) -> usize;
    fn is_empty(&self) -> bool;
}

MoveStore is a trait implemented by types which are able to store Moves inside themselves and are thus usable in move-generation methods in Position like Position::generate_moves_into<T>.

Required Methods

fn push(&mut self, m: Move)

push adds the given Move to the MoveStore.

fn len(&self) -> usize

len returns the number of Moves stored in the MoveStore.

fn is_empty(&self) -> bool

is_empty checks if no Moves are stored in the MoveStore.