Type Alias ataxx::MoveList

pub type MoveList = ArrayVec<Move, 256>;

MoveList is a basic implementation of MoveStore that is used to allow users to utilize move-generation methods without having to implement a MoveStore by themselves. It also has utility methods other than the MoveStore trait.

Aliased Type

struct MoveList { /* private fields */ }